% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_logo}
\alias{add_logo}
\title{Add a logo to your plot or image}
\usage{
add_logo(plot_path, logo_path, logo_position, logo_scale = 10)
}
\description{
This function places a logo at one of the corners of your image/plot
}
