% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{calculate_route}
\alias{calculate_route}
\title{Title}
\usage{
calculate_route(src, dst)
}
\arguments{
\item{src}{The location (vector of latitude, longitude) of the origin.}

\item{dst}{The location (vector of latitude, longitude) of the destination.}
}
\value{

}
\description{
Title
}
