% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulletgraph.R
\name{geom_bullet}
\alias{geom_bullet}
\title{Code to generate a bullet graph}
\usage{
geom_bullet(df, target, var_group, var_value, max_decimal = 0.85,
  high_decimal = 0.8, low_decimal = 0.6)
}
\description{
Code to generate a bullet graph
}
