% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{GET}
\alias{GET}
\title{Wrapper for httr::GET()}
\usage{
GET(url, query = list(), headers = list(), config = list(), retry = 5, ...)
}
\arguments{
\item{url}{URL to retrieve.}

\item{query}{Query parameters.}

\item{headers}{Header records.}

\item{config}{Additional configuration settings.}

\item{retry}{Number of times to retry request on failure.}

\item{...}{Further named parameters.}
}
\description{
Wrapper for httr::GET()
}
