% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{incident_details}
\alias{incident_details}
\title{Traffic incidents}
\usage{
incident_details(left, bottom, right, top)
}
\arguments{
\item{left}{Left edge of bounding box.}

\item{bottom}{Bottom edge of bounding box.}

\item{right}{Right edge of bounding box.}

\item{top}{Top edge of bounding box.}
}
\value{

}
\description{
Traffic incidents
}
\examples{
incidents <- incident_details(23.4, 37.9, 24.0, 38.2)
}
