% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{neg}
\alias{neg}
\title{A negative Function}
\usage{
neg(x)
}
\description{
This function allows you to create a negative number when it might
accidentally create an assignment operator.
}
\examples{
x < -5 # this actually assigns -5 to x, rather than
Checking whether x is less than -5.
neg()
}
\keyword{assignment}
\keyword{negative,}
