% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomtom.R
\name{set_api_key}
\alias{set_api_key}
\title{Set API key}
\usage{
set_api_key(api_key)
}
\arguments{
\item{api_key}{}
}
\value{

}
\description{
Set API key
}
\examples{
set_api_key("USf9S3AINkT4MhQZfz66ogYa3ZEGwPzZ")
}
