% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{std_err}
\alias{std_err}
\title{A standard error function}
\usage{
std_err(x)
}
\description{
This function allows you to calculate the standard error of a sample.
}
\examples{



}
\keyword{error}
\keyword{standard}
