% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{str_remove_right}
\alias{str_remove_right}
\title{A remove from end of string function}
\usage{
str_remove_right(x, n)
}
\description{
This function allows you to remove n characters starting from the end
of a string.
}
\examples{
Remove last 3 characters with str_remove_right(string, 3)
str_remove_right
}
\keyword{end,}
\keyword{remove}
\keyword{string,}
