% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{substr_right}
\alias{substr_right}
\title{A select from function}
\usage{
substr_right(x, n)
}
\description{
This function allows you to select characters starting from the end
of a string.
}
\examples{
Select last 3 characters with substr_right(string, 3)
substr_right()
}
\keyword{end}
\keyword{string,}
