% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tom_old.R
\name{theme_538}
\alias{theme_538}
\title{Personal theme inspired by ggthemes::theme_fivethirtyeight and fivethirtyeight.com}
\usage{
theme_538(base_size = 12, font = "Lato")
}
\description{
Theme inspired by the plots on
\href{ggthemes::theme_fivethirtyeight}{https://github.com/jrnold/ggthemes/blob/master/R/fivethirtyeight.R}.
}
