% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tom.R, R/theme_tom_old.R
\name{theme_tom}
\alias{theme_tom}
\title{Personal theme inspired by bbplot
https://github.com/bbc/bbplot/blob/master/R/bbc_style.R}
\usage{
theme_tom(base_size = 12, font = "Fira Mono")

theme_tom(base_size = 12, font = "Fira Mono")
}
\description{
Theme inspired by the plots on
\href{bbplot}{https://github.com/bbc/bbplot/blob/master/R/bbc_style.R}.

Theme inspired by the plots on
\href{ggthemes::theme_fivethirtyeight}{https://github.com/jrnold/ggthemes/blob/master/R/fivethirtyeight.R}.
}
