% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tomtom.R
\name{theme_tomtom}
\alias{theme_tomtom}
\title{My theme, adapted from 538}
\usage{
theme_tomtom(base_size = 13, base_family = "")
}
\description{
Theme inspired by the plots on
\href{fivethirtyeight.com}{http://fivethirtyeight.com}.
}
