% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tom.R
\name{tomtom_cols}
\alias{tomtom_cols}
\title{Function to extract tomtom colors as hex codes}
\usage{
tomtom_cols(...)
}
\arguments{
\item{...}{Character names of tomtom_colors}
}
\description{
Function to extract tomtom colors as hex codes
}
