% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_embed.R
\name{tweet_embed}
\alias{tweet_embed}
\title{Create markdown embed for a Tweet}
\usage{
tweet_embed(screen_name, status_id)
}
\arguments{
\item{screen_name}{character, screen name of the user}

\item{status_id}{character, status id}
}
\value{
character
}
\description{
The Twitter API GET call returns the tweet in markdown embedded format
}
\details{
Arguments to pass to the API call can be found \href{https://developer.twitter.com/en/docs/tweets/post-and-engage/api-reference/get-statuses-oembed}{here}
}
\examples{
name   <- 'kearneymw'
status <- '1087047171306856451'

tweet_embed(screen_name = name, status_id = status)

}
