% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tom.R
\name{theme_tom}
\alias{theme_tom}
\title{Personal theme inspired by ggthemes::theme_fivethirtyeight and fivethirtyeight.com}
\usage{
theme_tom(base_size = 12, font = "Titillium Web")
}
\description{
Theme inspired by the plots on
\href{ggthemes::theme_fivethirtyeight}{https://github.com/jrnold/ggthemes/blob/master/R/fivethirtyeight.R}.
}
