% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tom.R
\name{theme_tomtom}
\alias{theme_tomtom}
\title{Personal theme inspired by bbplot
https://github.com/bbc/bbplot/blob/master/R/bbc_style.R}
\usage{
theme_tomtom()
}
\description{
Theme inspired by the plots on
\href{bbplot}{https://github.com/bbc/bbplot/blob/master/R/bbc_style.R}.
}
